fix(runtime): Address CodeRabbit review feedback (OMN-2216)
- Add defensive AttributeError -> ImportError conversion in wiring.py
  getattr calls for consistent error handling
- Move callable check before is_class branch in wiring.py so all
  handler specs are validated uniformly
- Add clarifying comment in plugin.py initialize() explaining why the
  try block has no initialization work
- Extract duplicate test stub classes into shared conftest.py
- Replace fragile __qualname__ string matching with direct callable
  assertion in test_all_topics_use_dispatch_callback
